撤销加在 IndexWriter 外的锁, 由其自身处理

+++ hongs-serv-search/src/test/java/io/github/ihongs/test/TestWriter.java
@@ -1,6 +1,7 @@
 package io.github.ihongs.test;
 
 import io.github.ihongs.HongsException;
+import io.github.ihongs.dh.lucene.LuceneRecord;
 import io.github.ihongs.dh.search.SearchEntity;
 import io.github.ihongs.util.Synt;
 import java.io.File;
@@ -10,12 +11,33 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import static junit.framework.Assert.assertEquals;
+import junit.framework.TestCase;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 
-public class TestWriter {
+public class TestWriter extends TestCase {
 
-    //@Test
-    public void test() throws HongsException, InterruptedException {
+    @Before
+    @Override
+    public void setUp() {
+        // 删除测试的目录
+        delDir("test");
+        delDir("null");
+    }
+
+    @After
+    @Override
+    public void tearDown() {
+        // 删除测试的目录
+        delDir("test");
+        delDir("null");
+    }
+
+    @Test
+    public void testWriter() throws HongsException, InterruptedException, IOException {
+        if (1 == 1) return;
+        
         Map fields = Synt.mapOf(
             "id"   , Synt.mapOf(
                 "__name__", "id",
@@ -30,8 +52,8 @@ public class TestWriter {
                 "__type__", "datetime"
             )
         );
-        
-        final ExecutorService es = Executors.newFixedThreadPool(10);
+
+        final ExecutorService es = Executors.newFixedThreadPool(40);
         final Runnable rn = new Runnable() {
             @Override
             public void run() {
@@ -40,8 +62,8 @@ public class TestWriter {
                     SearchEntity se = new SearchEntity(fields, "test", "test");
                     se.begin ();
                     se.create(Synt.mapOf(
-                            "name" , "12ab",
-                            "time" , System.currentTimeMillis()
+                        "name" , "12ab",
+                        "time" , System.currentTimeMillis()
                     ));
                     se.commit();
                 }
@@ -50,34 +72,29 @@ public class TestWriter {
                 }
             }
         };
-        for (int i = 0; i < 10; i ++) {
+        for (int i = 0; i < 80; i ++) {
             es.execute(() -> rn.run());
         }
-        
+
         es.shutdown();
-        es.awaitTermination(10, TimeUnit.SECONDS);
-        
+        es.awaitTermination(20, TimeUnit.SECONDS);
+
         // 检查写入的内容
+    //  LuceneRecord se = new LuceneRecord(fields, "test", "test");
         SearchEntity se = new SearchEntity(fields, "test", "test");
-        assertEquals(10 , se.search( Synt.mapOf(), 0, 0 ).hits() );
-        try {
-            se.getWriter().close();
-        } catch ( IOException ex ) {
-            ex.printStackTrace ( );
-        }
-        
-        // 删除测试的目录
-        File dir;
-        dir = new File("test");
-        for(File file : dir.listFiles()){
-            file.delete();
-        }
-        dir.delete();
-        dir = new File("null");
-        for(File file : dir.listFiles()){
-            file.delete();
-        }
-        dir.delete();
+        assertEquals(80 , se.search( Synt.mapOf(), 0, 0 ).hits() );
+        se.getWriter().maybeMerge();
+        se.getWriter().deleteUnusedFiles();
+        se.getWriter().close();
     }
-    
+
+    private void delDir(String path) {
+        File dir = new File(path);
+        if ( dir.exists() ) {
+            for(File file : dir.listFiles()){
+                file.delete();
+            }   dir .delete();
+        }
+    }
+
 }
